I created a new helpers folder where I separated the functions that were together in data.js, the code is modularized.
