+++ src/index.html
@@ -13,21 +13,22 @@
 
 <body>
     <div class="container">
-        <section class="encabezado"></section>
-        <header class="header">
-            <div class="center">
-                <!-- LOGO -->
-                <div class="logo">
-                    <br><img src="imagenes/logofinal.png" alt="">
-                    <span class="brand">
+        <section class="encabezado">
+            <header class="header">
+                <div class="center">
+                    <!-- LOGO -->
+                    <div class="logo">
+                        <br><img src="imagenes/logofinal.png" alt="">
+                        <span class="brand">
                     <!--el titulo del sitio-->
                     ON-RANKEDS 
                      </span>
+                    </div>
+                    <div class="clearfix"></div>
                 </div>
-                <div class="clearfix"></div>
-            </div>
-        </header>
-        <!-- slider  -->
+            </header>
+        </section>
+
 
         <section id="top">
             <h2 class="title"> CHALLENGE <br> YOURSELF </h2>
@@ -63,13 +64,10 @@
                 </nav>
             </div>
         </section>
-
-        <!-- <button href="" id="jsonbotton" class="button">Mostrar api</button> -->
         <div id="cardsprueba" class="wrap"></div>
         <div id="onlyChampion"></div>
-
         <script type="text/javascript" src="main.js"></script>
-        <script type="text/javascript" src="data.js"></script>
+        <script type="module" src="data.js"></script>
     </div>
 </body>
